Moved AboutMacro to same location as other slash command macros. Added new chat function _slash_ that allows running slash commands in macros.
