+++ src/main/java/net/rptools/maptool/client/macro/impl/AboutMacro.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.client.functions;
+package net.rptools.maptool.client.macro.impl;
 
 import net.rptools.maptool.client.AppActions;
 import net.rptools.maptool.client.AppActions.ClientAction;
